requirements.txt
streamlit
reportlab==4.1.0
qrcode[pil]==7.4.2
pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pyzbar==0.1.9